+++ Bomberman/Game.cpp
@@ -92,8 +92,8 @@ Game::Game()
 	InitPlayers();
 	PlayerTextures();
 
-	players[0]->SpawnPlayer(sf::Vector2i(75, 74));
-	players[1]->SpawnPlayer(sf::Vector2i(675, 674));
+	players[0]->SpawnPlayer(sf::Vector2f(75.0f, 74.0f));
+	players[1]->SpawnPlayer(sf::Vector2f(675.0f, 674.0f));
 
 	players[0]->SetDirectionVisual();
 	players[1]->SetDirectionVisual();
